\name{define}
\alias{define}
\title{Define the generated code from generateInterface() in the current R session}
\description{
 This (relatively untested) function
 is intended to provide a way to read the code 
 generated by a call to \code{\link{generateInterface}}
 directly into an R session without writing it to a file.
 This allows us to dynamically create R wrappersto 
 DCOM interfaces without having the overhead of
 writing code to disk. 
 
 If one wants the interface to be persistent across R
 sessions, one should use \code{\link{writeCode}}.
}
\usage{
define(x, where, ...)
}
\arguments{
  \item{x}{the definition of the code generated from
   \code{\link{generateInterface}} (or other lower-level
    generation functions such as \code{generateClass}}
  \item{where}{the environment or index along the search path 
    in which to define the classes.}
  \item{\dots}{any additional arguments to govern the 
    processing of the definitions.}
}
\value{

}
\references{
 \url{http://msdn.microsoft.com/library}
 \url{http://www.omegahat.org/RWinTypeLibs}
}
\author{Duncan Temple Lang <duncan@wald.ucdavis.edu>}


\seealso{
 \code{\link{writeCode}}
 \code{\link{generateInterface}}
}
\examples{

}
\keyword{programming}
\keyword{interface}
\concept{meta-programming}


